# openapi.json
{"openapi":"3.0.0","paths":{"/g/booking/search":{"get":{"operationId":"BookingController_search","parameters":[{"name":"from","required":true,"in":"query","schema":{"type":"string"}},{"name":"to","required":true,"in":"query","schema":{"type":"string"}},{"name":"date","required":true,"in":"query","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponseDTO"}}}}},"tags":["Сервис бронирования"]}},"/g/booking/book":{"post":{"operationId":"BookingController_book","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDTO"}}}},"responses":{"200":{"description":""}},"tags":["Сервис бронирования"]}},"/g/booking/stand-queue":{"post":{"operationId":"BookingController_standQueue","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StandQueueDTO"}}}},"responses":{"200":{"description":""}},"tags":["Сервис бронирования"]}}},"info":{"title":"Dotg Gayway","description":"API specification for dotg-gateway","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"DetailedRoute":{"type":"object","properties":{"name":{"type":"string"},"num":{"type":"number"},"arrival":{"type":"string","example":"25.10.2024 11:07:00"},"departure":{"type":"string","example":"26.10.2024 13:09:00"}},"required":["name","num","arrival","departure"]},"Seat":{"type":"object","properties":{"seat_id":{"type":"number"},"seatNum":{"type":"string"},"block":{"type":"string"},"price":{"type":"number"},"bookingStatus":{"type":"string"}},"required":["seat_id","seatNum","block","price","bookingStatus"]},"Wagon":{"type":"object","properties":{"wagon_id":{"type":"number"},"type":{"type":"string"},"seats":{"type":"array","items":{"$ref":"#/components/schemas/Seat"}}},"required":["wagon_id","type","seats"]},"SearchResponseDTO":{"type":"object","properties":{"train_id":{"type":"number"},"global_route":{"type":"string","example":"Ростов-на-Дону->Москва"},"startpoint_departure":{"type":"string","example":"25.10.2024 11:07:00"},"endpoint_arrival":{"type":"string","example":"26.10.2024 13:09:00"},"detailed_route":{"type":"array","items":{"$ref":"#/components/schemas/DetailedRoute"}},"wagons_info":{"type":"array","items":{"$ref":"#/components/schemas/Wagon"}},"available_seats_count":{"type":"number"}},"required":["train_id","global_route","startpoint_departure","endpoint_arrival","detailed_route","wagons_info","available_seats_count"]},"BookDTO":{"type":"object","properties":{"trainId":{"type":"number"},"wagonId":{"type":"number"},"seatId":{"type":"number"}},"required":["trainId","wagonId","seatId"]},"StandQueueDTO":{"type":"object","properties":{"dateFrom":{"type":"string"},"dateTo":{"type":"string"},"from":{"type":"string"},"to":{"type":"string"},"priceFrom":{"type":"number"},"priceTo":{"type":"number"},"wagonType":{"type":"string"},"seatCount":{"type":"number"}},"required":["dateFrom","dateTo","from","to","priceFrom","priceTo","wagonType","seatCount"]}}}}